requirements-build.txt
pandas
openpyxl
scipy
matplotlib
numpy
jamdict
jamdict-data
jaconv
toml
regex
tabulate
pyyaml
